session_generation/requirements.txt
torch
transformers
sentence-transformers
tqdm
pandas
pyyaml
python-dotenv
huggingface_hub
